pyproject.toml
[project]
name = "jobfinder"
version = "3.0.0"
description = "CLI tool to discover relevant companies and open roles from your resume"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "anthropic>=0.42",
    "google-genai>=1.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "rich>=13.0",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0",
    "rapidfuzz>=3.0",
    "python-dateutil>=2.8",
    "supabase>=2.0",
    "pyjwt[crypto]>=2.0",
]

[project.optional-dependencies]
# Playwright: headless browser for rendering JS-heavy career pages (Tier 2, auto)
# After install: playwright install chromium
playwright = ["playwright>=1.40"]

# browser-use: LLM-driven agent for interactive career pages with pagination/filters (Tier 3, manual)
# Requires langchain bridge for whichever provider you use:
#   pip install "jobfinder[browser-use,langchain-anthropic]"    (Anthropic)
#   pip install "jobfinder[browser-use,langchain-google-genai]" (Gemini)
browser-use = ["browser-use>=0.1"]
langchain-anthropic = ["langchain-anthropic>=0.3"]
langchain-google-genai = ["langchain-google-genai>=2.0"]

# semantic: local embedding-based role filtering (one-time ~70MB ONNX model download)
#   pip install -e ".[semantic]"
semantic = ["fastembed>=0.3"]


# dev: test suite — pip install -e ".[dev]"
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "pytest-cov>=5.0",
    "httpx>=0.27",
]

[project.scripts]
jobfinder = "jobfinder.cli:cli"

[tool.setuptools.packages.find]
include = ["jobfinder*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
